test-commit

+++ Docker/clean_up_docker.sh
@@ -1,5 +1,14 @@
 #!/bin/bash
 
-docker stop $(docker ps -a -q)
-docker rm $(docker ps -a -q)
-docker rmi $(docker images -q)
+if [ `docker ps | wc -l` -gt 1 ]; then
+	docker stop $(docker ps -a -q)
+fi
+
+if [ `docker ps -a | wc -l` -gt 1 ]; then
+	docker rm $(docker ps -a -q)
+fi
+
+if [ `docker images | wc -l` -gt 1 ]; then
+	docker rmi $(docker images -q)
+fi
+
